+++ Assets/Scripts/skeletonArcher.cs
@@ -2,11 +2,11 @@
 using System.Collections;
 
 public class skeletonArcher : Tower {
-    private double initalDamage = 4;
-    private double initialFireRate = 2.0;
-    private double initialRange = 120;
+    private float initalDamage = 4;
+    private float initialFireRate = 2.0f;
+    private float initialRange = 120;
     private bool flying = true;
-    private double initialDoT = 2;
+    private float initialDoT = 2;
 
     // Use this for initialization
     void Start()
