+++ Assets/Scripts/undeadMage.cs
@@ -2,12 +2,12 @@
 using System.Collections;
 
 public class undeadMage : Tower {
-    private double initalDamage = 4;
-    private double initialFireRate = 1.0;
-    private double initialRange = 120;
+    private float initalDamage = 4;
+    private float initialFireRate = 1.0f;
+    private float initialRange = 120;
     private bool flying = true;
     private bool armorPen = true;
-    private double initialSlow = 0.7;
+    private float initialSlow = 0.7f;
 
     // Use this for initialization
     void Start()
